requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
SQLAlchemy==1.4.44
psycopg2-binary==2.9.5
avro-python3==1.10.2
python-multipart==0.0.5
